[TEST] Reorganice un poco mejor el codigo
